[BUGFIX] Fixes parsing of single forward slash
Bug has been introduced when I changed the regular
expression for comments to the one of the w3c. I
previously had this clause baked inside the comment
regex. It should be cleaner and better to read now.
A forward slash is now a valid char for css, but we
ensure that any comment is parsed before this data.